requirements.txt
discord-py
python-dotenv
pymongo
gradio_client
